web_application/lavidb/src/main/webapp/app/entities/database-version/route/database-version-routing-resolve.service.spec.ts
import { TestBed } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { ActivatedRouteSnapshot, ActivatedRoute, Router, convertToParamMap } from '@angular/router';
import { RouterTestingModule } from '@angular/router/testing';
import { of } from 'rxjs';

import { IDatabaseVersion } from '../database-version.model';
import { DatabaseVersionService } from '../service/database-version.service';

import { DatabaseVersionRoutingResolveService } from './database-version-routing-resolve.service';

describe('DatabaseVersion routing resolve service', () => {
  let mockRouter: Router;
  let mockActivatedRouteSnapshot: ActivatedRouteSnapshot;
  let routingResolveService: DatabaseVersionRoutingResolveService;
  let service: DatabaseVersionService;
  let resultDatabaseVersion: IDatabaseVersion | null | undefined;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule, RouterTestingModule.withRoutes([])],
      providers: [
        {
          provide: ActivatedRoute,
          useValue: {
            snapshot: {
              paramMap: convertToParamMap({}),
            },
          },
        },
      ],
    });
    mockRouter = TestBed.inject(Router);
    jest.spyOn(mockRouter, 'navigate').mockImplementation(() => Promise.resolve(true));
    mockActivatedRouteSnapshot = TestBed.inject(ActivatedRoute).snapshot;
    routingResolveService = TestBed.inject(DatabaseVersionRoutingResolveService);
    service = TestBed.inject(DatabaseVersionService);
    resultDatabaseVersion = undefined;
  });

  describe('resolve', () => {
    it('should return IDatabaseVersion returned by find', () => {
      // GIVEN
      service.find = jest.fn(id => of(new HttpResponse({ body: { id } })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultDatabaseVersion = result;
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultDatabaseVersion).toEqual({ id: 123 });
    });

    it('should return null if id is not provided', () => {
      // GIVEN
      service.find = jest.fn();
      mockActivatedRouteSnapshot.params = {};

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultDatabaseVersion = result;
      });

      // THEN
      expect(service.find).not.toBeCalled();
      expect(resultDatabaseVersion).toEqual(null);
    });

    it('should route to 404 page if data not found in server', () => {
      // GIVEN
      jest.spyOn(service, 'find').mockReturnValue(of(new HttpResponse<IDatabaseVersion>({ body: null })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultDatabaseVersion = result;
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultDatabaseVersion).toEqual(undefined);
      expect(mockRouter.navigate).toHaveBeenCalledWith(['404']);
    });
  });
});
